requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.42.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.6
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy>=1.24.3,<1.26
oauthlib==3.3.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scipy==1.13.1
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions>=4.6.0,<5.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
